+++ src/entrypoints/content/modules/links/utils.test.ts
@@ -28,13 +28,13 @@ describe(replacePlaceholders, () => {
     const url = 'https://example.com/team/{teamId}'
     const data = { teamId: null, seriesId: '456' }
 
-    expect(() => replacePlaceholders(url, data)).toThrowError('Missing data for placeholder {teamId}')
+    expect(() => replacePlaceholders(url, data)).toThrow('Missing data for placeholder {teamId}')
   })
 
   it('throws error when data is missing a key', () => {
     const url = 'https://example.com/team/{teamId}'
     const data = { seriesId: '456' }
 
-    expect(() => replacePlaceholders(url, data)).toThrowError('Missing data for placeholder {teamId}')
+    expect(() => replacePlaceholders(url, data)).toThrow('Missing data for placeholder {teamId}')
   })
 })
